pyproject.toml
[project]
name = "ti"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docling>=2.76.0",
    "langchain>=1.2.10",
    "langchain-gigachat-lc1==0.4.0b4",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.0.10",
    "numpy<2",
    "pandas<3",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pdftotext>=3.0.0",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "pytesseract>=0.3.13",
    "reportlab>=4.4.10",
    "rich>=14.3.3",
]
